backend/requirements.txt
aniso8601==9.0.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
virtualenv==20.16.3
Werkzeug==2.2.2
zipp==3.8.1
